requirements.txt
fastapi
uvicorn[standard]
aio-pika
pydantic
pydantic-settings
pymongo
requests
python-dotenv
loguru